Move online test repo to separate package
