+++ app/services/testdata/Phase1TestsResultsReceivedStatusGenerator.scala
@@ -19,7 +19,7 @@ package services.testdata
 import common.FutureEx
 import model.exchange.Phase1TestResultReady
 import repositories._
-import repositories.application.OnlineTestRepository
+import repositories.onlinetesting.OnlineTestRepository
 import services.onlinetesting.OnlineTestService
 import uk.gov.hmrc.play.http.HeaderCarrier
